Separated `data` package from web server.
