+++ data/dataset.py
@@ -1,6 +1,6 @@
+from data import create_cursor, commit
 import copy
 import jsonschema
-from web_server import data
 
 
 # JSON schema is used for validation of submitted datasets. BASE_JSON_SCHEMA
@@ -70,7 +70,7 @@ def create_from_dict(dictionary, author_id):
     """
     jsonschema.validate(dictionary, BASE_JSON_SCHEMA)
 
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         if "description" not in dictionary:
             dictionary["description"] = None
 
@@ -92,7 +92,7 @@ def create_from_dict(dictionary, author_id):
                                (cls_id, recording_mbid))
 
     # If anything bad happens above, it should just rollback by default.
-    data.commit()
+    commit()
 
     return dataset_id
 
@@ -101,7 +101,7 @@ def update(dataset_id, dictionary, author_id):
     # TODO(roman): Make author_id argument optional (keep old author if None).
     jsonschema.validate(dictionary, BASE_JSON_SCHEMA)
 
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         if "description" not in dictionary:
             dictionary["description"] = None
 
@@ -126,7 +126,7 @@ def update(dataset_id, dictionary, author_id):
                                (cls_id, recording_mbid))
 
     # If anything bad happens above, it should just rollback by default.
-    data.commit()
+    commit()
 
 
 def get(id):
@@ -136,7 +136,7 @@ def get(id):
         Dictionary with dataset details if it has been found, None
         otherwise.
     """
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         cursor.execute(
             "SELECT id, name, description, author, created, public "
             "FROM dataset "
@@ -152,7 +152,7 @@ def get(id):
 
 
 def _get_classes(dataset_id):
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         cursor.execute(
             "SELECT id, name, description "
             "FROM dataset_class "
@@ -169,7 +169,7 @@ def _get_classes(dataset_id):
 
 
 def _get_recordings_in_class(class_id):
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         cursor.execute("SELECT mbid FROM dataset_class_member WHERE class = %s",
                        (class_id,))
         rows = cursor.fetchall()
@@ -185,7 +185,7 @@ def get_by_user_id(user_id, public_only=True):
     Returns:
         List of dictionaries with dataset details.
     """
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         where = "WHERE author = %s"
         if public_only:
             where += " AND public = TRUE"
@@ -197,6 +197,6 @@ def get_by_user_id(user_id, public_only=True):
 
 def delete(id):
     """Delete dataset with a specified ID."""
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         cursor.execute("DELETE FROM dataset WHERE id = %s", (str(id),))
-    data.commit()
+    commit()
