+++ data/dump.py
@@ -10,8 +10,8 @@ include information from the previous dumps).
 from __future__ import print_function
 from collections import defaultdict
 from datetime import datetime
-from web_server import data
-from web_server.utils import create_path
+from data import create_cursor, commit
+from data.utils import create_path
 import web_server
 import subprocess
 import tempfile
@@ -144,7 +144,7 @@ def _copy_tables(location, start_time=None, end_time=None):
         else:
             return ""
 
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
 
         # lowlevel
         with open(os.path.join(location, "lowlevel"), "w") as f:
@@ -185,7 +185,7 @@ def import_db_dump(archive_path):
 
     table_names = _TABLES.keys()
 
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
 
         with tarfile.open(fileobj=pxz.stdout, mode="r|") as tar:
             for member in tar:
@@ -207,7 +207,7 @@ def import_db_dump(archive_path):
                         cursor.copy_from(tar.extractfile(member), '"%s"' % file_name,
                                          columns=_TABLES[file_name])
 
-    data.commit()
+    commit()
     pxz.stdout.close()
 
 
@@ -237,7 +237,7 @@ def dump_lowlevel_json(location, incremental=False, dump_id=None):
     archive_path = os.path.join(location, archive_name + ".tar.bz2")
     with tarfile.open(archive_path, "w:bz2") as tar:
 
-        with data.create_cursor() as cursor:
+        with create_cursor() as cursor:
 
             mbid_occurences = defaultdict(int)
 
@@ -264,7 +264,7 @@ def dump_lowlevel_json(location, incremental=False, dump_id=None):
                 where = ""
             cursor.execute("SELECT id FROM lowlevel ll %s ORDER BY mbid" % where)
 
-            with data.create_cursor() as cursor_inner:
+            with create_cursor() as cursor_inner:
                 temp_dir = tempfile.mkdtemp()
 
                 dumped_count = 0
@@ -333,7 +333,7 @@ def dump_highlevel_json(location, incremental=False, dump_id=None):
     archive_path = os.path.join(location, archive_name + ".tar.bz2")
     with tarfile.open(archive_path, "w:bz2") as tar:
 
-        with data.create_cursor() as cursor:
+        with create_cursor() as cursor:
             mbid_occurences = defaultdict(int)
 
             # Need to count how many duplicate MBIDs are there before start_time
@@ -362,7 +362,7 @@ def dump_highlevel_json(location, incremental=False, dump_id=None):
                               WHERE hl.data = hlj.id %s
                               ORDER BY mbid""" % where)
 
-            with data.create_cursor() as cursor_inner:
+            with create_cursor() as cursor_inner:
                 temp_dir = tempfile.mkdtemp()
 
                 dumped_count = 0
@@ -412,7 +412,7 @@ def list_incremental_dumps():
         List of (id, created) pairs ordered by dump identifier, or None if
         there are no incremental dumps yet.
     """
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         cursor.execute("SELECT id, created FROM incremental_dumps ORDER BY id DESC")
         return cursor.fetchall()
 
@@ -449,7 +449,7 @@ def _any_new_data(from_time):
         True if there is new data in one of tables that support incremental
         dumps, False if there is no new data there.
     """
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         cursor.execute("SELECT count(*) FROM lowlevel WHERE submitted > %s", (from_time,))
         lowlevel_count = cursor.fetchone()[0]
         cursor.execute("SELECT count(*) FROM highlevel WHERE submitted > %s", (from_time,))
@@ -459,16 +459,16 @@ def _any_new_data(from_time):
 
 def _create_new_inc_dump_record():
     """Creates new record for incremental dump and returns its ID and creation time."""
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         cursor.execute("INSERT INTO incremental_dumps (created) VALUES (now()) RETURNING id, created")
-        data.commit()
+        commit()
         row = cursor.fetchone()
     print("Created new incremental dump record (ID: %s)." % row[0])
     return row
 
 
 def _get_incremental_dump_timestamp(dump_id=None):
-    with data.create_cursor() as cursor:
+    with create_cursor() as cursor:
         if dump_id:
             cursor.execute("SELECT created FROM incremental_dumps WHERE id = %s", (dump_id,))
         else:
