+++ data/user_test.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-from web_server.testing import ServerTestCase
-from web_server.data import user as user_data
+from data.testing import DataTestCase
+from data import user as user_data
 
 
-class UserTestCase(ServerTestCase):
+class UserTestCase(DataTestCase):
 
     def test_create(self):
         user_id = user_data.create("fuzzy_dunlop")
@@ -27,7 +27,7 @@ class UserTestCase(ServerTestCase):
         user_id = user_data.create(musicbrainz_id)
         user = user_data.get_by_mb_id(musicbrainz_id)
         self.assertIsNotNone(user)
-        self.assertEqual(user.id, user_id)
+        self.assertEqual(user["id"], user_id)
 
     def test_get_or_create(self):
         musicbrainz_id = "User"
@@ -35,7 +35,7 @@ class UserTestCase(ServerTestCase):
         self.assertIsNotNone(user)
 
         same_user = user_data.get_or_create(musicbrainz_id)
-        self.assertEqual(user.id, same_user.id)
+        self.assertEqual(user["id"], same_user["id"])
 
         # Testing Unicode
         musicbrainz_id = u"Пользователь"
@@ -43,4 +43,4 @@ class UserTestCase(ServerTestCase):
         self.assertIsNotNone(user)
 
         same_user = user_data.get_or_create(musicbrainz_id)
-        self.assertEqual(user.id, same_user.id)
+        self.assertEqual(user["id"], same_user["id"])
