+++ web_server/external/musicbrainz.py
@@ -1,6 +1,6 @@
 import musicbrainzngs
 from musicbrainzngs.musicbrainz import ResponseError
-from web_server import cache
+from data import cache
 
 CACHE_TIMEOUT = 86400  # 1 day
