+++ web_server/views/api_test.py
@@ -1,4 +1,6 @@
-from web_server.testing import ServerTestCase, TEST_DATA_PATH
+from __future__ import absolute_import
+from web_server.testing import ServerTestCase
+from data.testing import TEST_DATA_PATH
 from flask import url_for
 import os
