+++ web_server/views/stats.py
@@ -1,5 +1,6 @@
+from __future__ import absolute_import
 from flask import Blueprint, Response, render_template
-from web_server.data.stats import get_statistics_data
+from data.stats import get_statistics_data
 from operator import itemgetter
 import json
